+++ src/parser/enrichers/monster/Stirge/Proboscis.mjs
@@ -1,0 +1,61 @@
+/* eslint-disable class-methods-use-this */
+// import { utils } from "../../../../lib/_module.mjs";
+import DDBEnricherData from "../../data/DDBEnricherData.mjs";
+
+export default class Proboscis extends DDBEnricherData {
+  get type() {
+    return "attack";
+  }
+
+  get activity() {
+    return {
+      targetType: "creature",
+      noTemplate: true,
+      data: {
+        damage: {
+          parts: [
+            DDBEnricherData.basicDamagePart({
+              number: 1,
+              denomination: 6,
+              bonus: "@mod",
+              types: ["piercing"],
+            }),
+          ],
+        },
+      },
+    };
+  }
+
+  get additionalActivities() {
+    return [
+      {
+        constructor: {
+          name: "Attached Damage",
+          type: "damage",
+        },
+        build: {
+          generateDamage: true,
+          generateTarget: true,
+        },
+        overrides: {
+          targetType: "creature",
+          activationType: "special",
+          activationCondition: "Start of the Stirge's turn",
+          noTemplate: true,
+          data: {
+            damage: {
+              parts: [
+                DDBEnricherData.basicDamagePart({
+                  number: 2,
+                  denomination: 4,
+                  types: ["necrotic"],
+                }),
+              ],
+            },
+          },
+        },
+      },
+    ];
+  }
+
+}