+++ src/parser/enrichers/monster/Stirge/_module.mjs
@@ -1,0 +1,1 @@
+export { default as Proboscis } from "./Proboscis.mjs";