+++ src/parser/enrichers/monster/_module.mjs
@@ -21,3 +21,4 @@ export * as SummonAnimatedObject from "./SummonAnimatedObject/_module.mjs";
 export * as SummonConstruct from "./SummonConstruct/_module.mjs";
 export * as Aboleth from "./Aboleth/_module.mjs";
 export * as EldritchCannon from "./EldritchCannon/_module.mjs";
+export * as Stirge from "./Stirge/_module.mjs";