+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Next Up
+
+- Stirge 2024 damage parsing corrections @ekcameron
+
+
 # 6.1.0
 
 - New UI for DDB Character Importer, DDB Muncher and DDB Encounter Muncher! (migrated to Application V2). Encounter munch is now part of the regular muncher.